feat: flow and storages

+++ lib-node6-dev/pool/Process.js
@@ -44,6 +44,7 @@ class Process {
   }
 
   kill() {
+    if (!this.childProcess) return;
     this.childProcess.kill();
     delete this.childProcess;
   }
@@ -55,6 +56,10 @@ class Process {
   startTeam(team) {
     _assert(team, _types.TeamType, 'team');
 
+    if (!this.childProcess) {
+      throw new Error('Cannot start a new team in a killed process');
+    }
+
     this.teams.set(team.id, team);
     this.pool.teamsToProcess.set(team.id, this);
 
@@ -63,6 +68,12 @@ class Process {
 
   killTeam(team, killProcessIfEmpty = true) {
     _assert(team, _types.TeamType, 'team');
+
+    _assert(killProcessIfEmpty, _tcombForked2.default.Boolean, 'killProcessIfEmpty');
+
+    if (!this.childProcess) {
+      throw new Error('Cannot kill a team in a killed process');
+    }
 
     this.childProcess.send({ type: 'remove', teamId: team.id });
     this.teams.delete(team.id);
@@ -85,6 +96,10 @@ class Process {
     _assert(teamId, _tcombForked2.default.Number, 'teamId');
 
     _assert(data, _tcombForked2.default.Object, 'data');
+
+    if (!this.childProcess) {
+      throw new Error('Cannot send a message in a killed process');
+    }
 
     this.childProcess.send(_extends({ type: 'message', teamId }, data));
   }
